backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pydantic
python-multipart
openai
